+++ extensions/cronet/src/main/java/com/google/android/exoplayer2/ext/cronet/CronetDataSource.java
@@ -25,9 +25,12 @@ import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.ExoPlayerLibraryInfo;
 import com.google.android.exoplayer2.upstream.BaseDataSource;
+import com.google.android.exoplayer2.upstream.DataSource;
 import com.google.android.exoplayer2.upstream.DataSourceException;
 import com.google.android.exoplayer2.upstream.DataSpec;
+import com.google.android.exoplayer2.upstream.DefaultHttpDataSource;
 import com.google.android.exoplayer2.upstream.HttpDataSource;
+import com.google.android.exoplayer2.upstream.TransferListener;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Clock;
 import com.google.android.exoplayer2.util.ConditionVariable;
@@ -65,6 +68,186 @@ import org.chromium.net.UrlResponseInfo;
  */
 public class CronetDataSource extends BaseDataSource implements HttpDataSource {
 
+  static {
+    ExoPlayerLibraryInfo.registerModule("goog.exo.cronet");
+  }
+
+  /** {@link DataSource.Factory} for {@link CronetDataSource} instances. */
+  public static final class Factory implements HttpDataSource.Factory {
+
+    private final CronetEngineWrapper cronetEngineWrapper;
+    private final Executor executor;
+    private final RequestProperties defaultRequestProperties;
+    private final DefaultHttpDataSource.Factory internalFallbackFactory;
+
+    @Nullable private HttpDataSource.Factory fallbackFactory;
+    @Nullable private Predicate<String> contentTypePredicate;
+    @Nullable private TransferListener transferListener;
+    private int connectTimeoutMs;
+    private int readTimeoutMs;
+    private boolean resetTimeoutOnRedirects;
+    private boolean handleSetCookieRequests;
+
+    /**
+     * Creates an instance.
+     *
+     * @param cronetEngineWrapper A {@link CronetEngineWrapper}.
+     * @param executor The {@link java.util.concurrent.Executor} that will handle responses. This
+     *     may be a direct executor (i.e. executes tasks on the calling thread) in order to avoid a
+     *     thread hop from Cronet's internal network thread to the response handling thread.
+     *     However, to avoid slowing down overall network performance, care must be taken to make
+     *     sure response handling is a fast operation when using a direct executor.
+     */
+    public Factory(CronetEngineWrapper cronetEngineWrapper, Executor executor) {
+      this.cronetEngineWrapper = cronetEngineWrapper;
+      this.executor = executor;
+      defaultRequestProperties = new RequestProperties();
+      internalFallbackFactory = new DefaultHttpDataSource.Factory();
+      connectTimeoutMs = DEFAULT_CONNECT_TIMEOUT_MILLIS;
+      readTimeoutMs = DEFAULT_READ_TIMEOUT_MILLIS;
+    }
+
+    /** @deprecated Use {@link #setDefaultRequestProperties(Map)} instead. */
+    @Deprecated
+    @Override
+    public final RequestProperties getDefaultRequestProperties() {
+      return defaultRequestProperties;
+    }
+
+    @Override
+    public final Factory setDefaultRequestProperties(Map<String, String> defaultRequestProperties) {
+      this.defaultRequestProperties.clearAndSet(defaultRequestProperties);
+      internalFallbackFactory.setDefaultRequestProperties(defaultRequestProperties);
+      return this;
+    }
+
+    /**
+     * Sets the connect timeout, in milliseconds.
+     *
+     * <p>The default is {@link CronetDataSource#DEFAULT_CONNECT_TIMEOUT_MILLIS}.
+     *
+     * @param connectTimeoutMs The connect timeout, in milliseconds, that will be used.
+     * @return This factory.
+     */
+    public Factory setConnectionTimeoutMs(int connectTimeoutMs) {
+      this.connectTimeoutMs = connectTimeoutMs;
+      internalFallbackFactory.setConnectTimeoutMs(connectTimeoutMs);
+      return this;
+    }
+
+    /**
+     * Sets whether the connect timeout is reset when a redirect occurs.
+     *
+     * <p>The default is {@code false}.
+     *
+     * @param resetTimeoutOnRedirects Whether the connect timeout is reset when a redirect occurs.
+     * @return This factory.
+     */
+    public Factory setResetTimeoutOnRedirects(boolean resetTimeoutOnRedirects) {
+      this.resetTimeoutOnRedirects = resetTimeoutOnRedirects;
+      return this;
+    }
+
+    /**
+     * Sets whether "Set-Cookie" requests on redirect should be forwarded to the redirect url in the
+     * "Cookie" header.
+     *
+     * <p>The default is {@code false}.
+     *
+     * @param handleSetCookieRequests Whether "Set-Cookie" requests on redirect should be forwarded
+     *     to the redirect url in the "Cookie" header.
+     * @return This factory.
+     */
+    public Factory setHandleSetCookieRequests(boolean handleSetCookieRequests) {
+      this.handleSetCookieRequests = handleSetCookieRequests;
+      return this;
+    }
+
+    /**
+     * Sets the read timeout, in milliseconds.
+     *
+     * <p>The default is {@link CronetDataSource#DEFAULT_READ_TIMEOUT_MILLIS}.
+     *
+     * @param readTimeoutMs The connect timeout, in milliseconds, that will be used.
+     * @return This factory.
+     */
+    public Factory setReadTimeoutMs(int readTimeoutMs) {
+      this.readTimeoutMs = readTimeoutMs;
+      internalFallbackFactory.setReadTimeoutMs(readTimeoutMs);
+      return this;
+    }
+
+    /**
+     * Sets a content type {@link Predicate}. If a content type is rejected by the predicate then a
+     * {@link HttpDataSource.InvalidContentTypeException} is thrown from {@link #open(DataSpec)}.
+     *
+     * <p>The default is {@code null}.
+     *
+     * @param contentTypePredicate The content type {@link Predicate}, or {@code null} to clear a
+     *     predicate that was previously set.
+     * @return This factory.
+     */
+    public Factory setContentTypePredicate(@Nullable Predicate<String> contentTypePredicate) {
+      this.contentTypePredicate = contentTypePredicate;
+      internalFallbackFactory.setContentTypePredicate(contentTypePredicate);
+      return this;
+    }
+
+    /**
+     * Sets the {@link TransferListener} that will be used.
+     *
+     * <p>The default is {@code null}.
+     *
+     * <p>See {@link DataSource#addTransferListener(TransferListener)}.
+     *
+     * @param transferListener The listener that will be used.
+     * @return This factory.
+     */
+    public Factory setTransferListener(@Nullable TransferListener transferListener) {
+      this.transferListener = transferListener;
+      internalFallbackFactory.setTransferListener(transferListener);
+      return this;
+    }
+
+    /**
+     * Sets the fallback {@link HttpDataSource.Factory} that is used as a fallback if the {@link
+     * CronetEngineWrapper} fails to provide a {@link CronetEngine}.
+     *
+     * <p>By default a {@link DefaultHttpDataSource} is used as fallback factory.
+     *
+     * @param fallbackFactory The fallback factory that will be used.
+     * @return This factory.
+     */
+    public Factory setFallbackFactory(@Nullable HttpDataSource.Factory fallbackFactory) {
+      this.fallbackFactory = fallbackFactory;
+      return this;
+    }
+
+    @Override
+    public HttpDataSource createDataSource() {
+      @Nullable CronetEngine cronetEngine = cronetEngineWrapper.getCronetEngine();
+      if (cronetEngine == null) {
+        return (fallbackFactory != null)
+            ? fallbackFactory.createDataSource()
+            : internalFallbackFactory.createDataSource();
+      }
+      CronetDataSource dataSource =
+          new CronetDataSource(
+              cronetEngine,
+              executor,
+              connectTimeoutMs,
+              readTimeoutMs,
+              resetTimeoutOnRedirects,
+              handleSetCookieRequests,
+              defaultRequestProperties,
+              contentTypePredicate);
+      if (transferListener != null) {
+        dataSource.addTransferListener(transferListener);
+      }
+      return dataSource;
+    }
+  }
+
   /**
    * Thrown when an error is encountered when trying to open a {@link CronetDataSource}.
    */
@@ -85,20 +268,11 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
       super(errorMessage, dataSpec, TYPE_OPEN);
       this.cronetConnectionStatus = cronetConnectionStatus;
     }
-
   }
 
-  static {
-    ExoPlayerLibraryInfo.registerModule("goog.exo.cronet");
-  }
-
-  /**
-   * The default connection timeout, in milliseconds.
-   */
+  /** The default connection timeout, in milliseconds. */
   public static final int DEFAULT_CONNECT_TIMEOUT_MILLIS = 8 * 1000;
-  /**
-   * The default read timeout, in milliseconds.
-   */
+  /** The default read timeout, in milliseconds. */
   public static final int DEFAULT_READ_TIMEOUT_MILLIS = 8 * 1000;
 
   /* package */ final UrlRequest.Callback urlRequestCallback;
@@ -149,16 +323,9 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
 
   private volatile long currentConnectTimeoutMs;
 
-  /**
-   * Creates an instance.
-   *
-   * @param cronetEngine A CronetEngine.
-   * @param executor The {@link java.util.concurrent.Executor} that will handle responses. This may
-   *     be a direct executor (i.e. executes tasks on the calling thread) in order to avoid a thread
-   *     hop from Cronet's internal network thread to the response handling thread. However, to
-   *     avoid slowing down overall network performance, care must be taken to make sure response
-   *     handling is a fast operation when using a direct executor.
-   */
+  /** @deprecated Use {@link CronetDataSource.Factory} instead. */
+  @SuppressWarnings("deprecation")
+  @Deprecated
   public CronetDataSource(CronetEngine cronetEngine, Executor executor) {
     this(
         cronetEngine,
@@ -169,21 +336,8 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
         /* defaultRequestProperties= */ null);
   }
 
-  /**
-   * Creates an instance.
-   *
-   * @param cronetEngine A CronetEngine.
-   * @param executor The {@link java.util.concurrent.Executor} that will handle responses. This may
-   *     be a direct executor (i.e. executes tasks on the calling thread) in order to avoid a thread
-   *     hop from Cronet's internal network thread to the response handling thread. However, to
-   *     avoid slowing down overall network performance, care must be taken to make sure response
-   *     handling is a fast operation when using a direct executor.
-   * @param connectTimeoutMs The connection timeout, in milliseconds.
-   * @param readTimeoutMs The read timeout, in milliseconds.
-   * @param resetTimeoutOnRedirects Whether the connect timeout is reset when a redirect occurs.
-   * @param defaultRequestProperties Optional default {@link RequestProperties} to be sent to the
-   *     server as HTTP headers on every request.
-   */
+  /** @deprecated Use {@link CronetDataSource.Factory} instead. */
+  @Deprecated
   public CronetDataSource(
       CronetEngine cronetEngine,
       Executor executor,
@@ -197,28 +351,13 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
         connectTimeoutMs,
         readTimeoutMs,
         resetTimeoutOnRedirects,
-        Clock.DEFAULT,
+        /* handleSetCookieRequests= */ false,
         defaultRequestProperties,
-        /* handleSetCookieRequests= */ false);
+        /* contentTypePredicate= */ null);
   }
 
-  /**
-   * Creates an instance.
-   *
-   * @param cronetEngine A CronetEngine.
-   * @param executor The {@link java.util.concurrent.Executor} that will handle responses. This may
-   *     be a direct executor (i.e. executes tasks on the calling thread) in order to avoid a thread
-   *     hop from Cronet's internal network thread to the response handling thread. However, to
-   *     avoid slowing down overall network performance, care must be taken to make sure response
-   *     handling is a fast operation when using a direct executor.
-   * @param connectTimeoutMs The connection timeout, in milliseconds.
-   * @param readTimeoutMs The read timeout, in milliseconds.
-   * @param resetTimeoutOnRedirects Whether the connect timeout is reset when a redirect occurs.
-   * @param defaultRequestProperties Optional default {@link RequestProperties} to be sent to the
-   *     server as HTTP headers on every request.
-   * @param handleSetCookieRequests Whether "Set-Cookie" requests on redirect should be forwarded to
-   *     the redirect url in the "Cookie" header.
-   */
+  /** @deprecated Use {@link CronetDataSource.Factory} instead. */
+  @Deprecated
   public CronetDataSource(
       CronetEngine cronetEngine,
       Executor executor,
@@ -233,26 +372,12 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
         connectTimeoutMs,
         readTimeoutMs,
         resetTimeoutOnRedirects,
-        Clock.DEFAULT,
+        handleSetCookieRequests,
         defaultRequestProperties,
-        handleSetCookieRequests);
+        /* contentTypePredicate= */ null);
   }
 
-  /**
-   * Creates an instance.
-   *
-   * @param cronetEngine A CronetEngine.
-   * @param executor The {@link java.util.concurrent.Executor} that will handle responses. This may
-   *     be a direct executor (i.e. executes tasks on the calling thread) in order to avoid a thread
-   *     hop from Cronet's internal network thread to the response handling thread. However, to
-   *     avoid slowing down overall network performance, care must be taken to make sure response
-   *     handling is a fast operation when using a direct executor.
-   * @param contentTypePredicate An optional {@link Predicate}. If a content type is rejected by the
-   *     predicate then an {@link InvalidContentTypeException} is thrown from {@link
-   *     #open(DataSpec)}.
-   * @deprecated Use {@link #CronetDataSource(CronetEngine, Executor)} and {@link
-   *     #setContentTypePredicate(Predicate)}.
-   */
+  /** @deprecated Use {@link CronetDataSource.Factory} instead. */
   @SuppressWarnings("deprecation")
   @Deprecated
   public CronetDataSource(
@@ -269,26 +394,7 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
         /* defaultRequestProperties= */ null);
   }
 
-  /**
-   * Creates an instance.
-   *
-   * @param cronetEngine A CronetEngine.
-   * @param executor The {@link java.util.concurrent.Executor} that will handle responses. This may
-   *     be a direct executor (i.e. executes tasks on the calling thread) in order to avoid a thread
-   *     hop from Cronet's internal network thread to the response handling thread. However, to
-   *     avoid slowing down overall network performance, care must be taken to make sure response
-   *     handling is a fast operation when using a direct executor.
-   * @param contentTypePredicate An optional {@link Predicate}. If a content type is rejected by the
-   *     predicate then an {@link InvalidContentTypeException} is thrown from {@link
-   *     #open(DataSpec)}.
-   * @param connectTimeoutMs The connection timeout, in milliseconds.
-   * @param readTimeoutMs The read timeout, in milliseconds.
-   * @param resetTimeoutOnRedirects Whether the connect timeout is reset when a redirect occurs.
-   * @param defaultRequestProperties Optional default {@link RequestProperties} to be sent to the
-   *     server as HTTP headers on every request.
-   * @deprecated Use {@link #CronetDataSource(CronetEngine, Executor, int, int, boolean,
-   *     RequestProperties)} and {@link #setContentTypePredicate(Predicate)}.
-   */
+  /** @deprecated Use {@link CronetDataSource.Factory} instead. */
   @SuppressWarnings("deprecation")
   @Deprecated
   public CronetDataSource(
@@ -310,28 +416,7 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
         /* handleSetCookieRequests= */ false);
   }
 
-  /**
-   * Creates an instance.
-   *
-   * @param cronetEngine A CronetEngine.
-   * @param executor The {@link java.util.concurrent.Executor} that will handle responses. This may
-   *     be a direct executor (i.e. executes tasks on the calling thread) in order to avoid a thread
-   *     hop from Cronet's internal network thread to the response handling thread. However, to
-   *     avoid slowing down overall network performance, care must be taken to make sure response
-   *     handling is a fast operation when using a direct executor.
-   * @param contentTypePredicate An optional {@link Predicate}. If a content type is rejected by the
-   *     predicate then an {@link InvalidContentTypeException} is thrown from {@link
-   *     #open(DataSpec)}.
-   * @param connectTimeoutMs The connection timeout, in milliseconds.
-   * @param readTimeoutMs The read timeout, in milliseconds.
-   * @param resetTimeoutOnRedirects Whether the connect timeout is reset when a redirect occurs.
-   * @param defaultRequestProperties Optional default {@link RequestProperties} to be sent to the
-   *     server as HTTP headers on every request.
-   * @param handleSetCookieRequests Whether "Set-Cookie" requests on redirect should be forwarded to
-   *     the redirect url in the "Cookie" header.
-   * @deprecated Use {@link #CronetDataSource(CronetEngine, Executor, int, int, boolean,
-   *     RequestProperties, boolean)} and {@link #setContentTypePredicate(Predicate)}.
-   */
+  /** @deprecated Use {@link CronetDataSource.Factory} instead. */
   @Deprecated
   public CronetDataSource(
       CronetEngine cronetEngine,
@@ -348,31 +433,31 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
         connectTimeoutMs,
         readTimeoutMs,
         resetTimeoutOnRedirects,
-        Clock.DEFAULT,
+        handleSetCookieRequests,
         defaultRequestProperties,
-        handleSetCookieRequests);
-    this.contentTypePredicate = contentTypePredicate;
+        contentTypePredicate);
   }
 
-  /* package */ CronetDataSource(
+  private CronetDataSource(
       CronetEngine cronetEngine,
       Executor executor,
       int connectTimeoutMs,
       int readTimeoutMs,
       boolean resetTimeoutOnRedirects,
-      Clock clock,
+      boolean handleSetCookieRequests,
       @Nullable RequestProperties defaultRequestProperties,
-      boolean handleSetCookieRequests) {
+      @Nullable Predicate<String> contentTypePredicate) {
     super(/* isNetwork= */ true);
-    this.urlRequestCallback = new UrlRequestCallback();
     this.cronetEngine = Assertions.checkNotNull(cronetEngine);
     this.executor = Assertions.checkNotNull(executor);
     this.connectTimeoutMs = connectTimeoutMs;
     this.readTimeoutMs = readTimeoutMs;
     this.resetTimeoutOnRedirects = resetTimeoutOnRedirects;
-    this.clock = Assertions.checkNotNull(clock);
     this.defaultRequestProperties = defaultRequestProperties;
     this.handleSetCookieRequests = handleSetCookieRequests;
+    this.contentTypePredicate = contentTypePredicate;
+    clock = Clock.DEFAULT;
+    urlRequestCallback = new UrlRequestCallback();
     requestProperties = new RequestProperties();
     operation = new ConditionVariable();
   }
@@ -384,6 +469,7 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
    * @param contentTypePredicate The content type {@link Predicate}, or {@code null} to clear a
    *     predicate that was previously set.
    */
+  @Deprecated
   public void setContentTypePredicate(@Nullable Predicate<String> contentTypePredicate) {
     this.contentTypePredicate = contentTypePredicate;
   }